seq change to integer instead of short
git-svn-id: https://everest.teltel.com:8443/repos/misc/web/nnvmso@1957 a0a5f996-2bb4-401c-8778-395c3c01700f

+++ src/com/nnvmso/service/TranscodingService.java
@@ -269,7 +269,7 @@ public class TranscodingService {
 		}
 		program.setPublic(true);
 		if (item.getSortId() != null)
-			program.setSeq(Short.valueOf(item.getSortId()));
+			program.setSeq(Integer.valueOf(item.getSortId()));
 		if (isNew) {
 			programMngr.create(channel, program);
 		} else {
